ios and android share index
