+++ index.ios.js
@@ -1,63 +1,4 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
- */
-
-import React, {Component} from 'react';
-import {
-    AppRegistry,
-    Alert,
-} from 'react-native';
-
-import {EventEmitter} from 'fbemitter';
-
-import Settings from './app/views/Settings';
-import Home from './app/views/Home';
-
-let _emitter = new EventEmitter();
-import DrawerLayout from './app/layouts/DrawerLayout';
-
-
-export default class ImadaApp extends Component {
-    componentDidMount() {
-        _emitter.addListener('back', () => {
-            this._navigator.pop();
-        });
-    }
-
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            view: <Home/>,
-        };
-    }
-
-    render() {
-        return (
-            <DrawerLayout
-                rightButtonPress={ () => this.navBarRightButtonPressed }
-                renderContent={ (route, navigator) => this.renderSceneView(route, navigator) }
-            />
-        );
-    }
-
-    renderSceneView(route, navigator) {
-        switch (route.id) {
-            case 'Home':
-                return ( <Home navigator={navigator}/> );
-
-            case 'Settings':
-                return ( <Settings navigator={navigator}/>);
-        }
-    }
-
-    navBarRightButtonPressed() {
-        Alert.alert('About', 'Created by Unknown Host.\n\nDevelopers:\n' +
-            'Christian Moeslund, chmoe13@student.sdu.dk\nJonatan Møller jogoe12@student.sdu.dk\n\n' +
-            'Other duderinos:\nAndreas Munk Jensen, Ehsanullah Ekhlas, Erik Zijdemans, Søren Anthony')
-    }
-}
+import React, { AppRegistry } from 'react-native';
+import ImadaApp from './app/index';
 
 AppRegistry.registerComponent('testproject', () => ImadaApp);
